Create DoorWall.cs
The code which creates an exterior wall (4) and place my door family into it.  Additionally I have my excel which gives the point coordinates where my door has to be placed and it additionally gives us an option that in which piece of wall among these 4 exterior wall my door has to be embedded.